# server/api.gen.go
// Package server provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.7.0 DO NOT EDIT.
package server

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Detail A human-readable explanation specific to this occurrence.
	Detail *string `json:"detail,omitempty"`

	// Instance A URI reference for this specific occurrence (e.g. trace ID).
	Instance *string `json:"instance,omitempty"`

	// Status The HTTP status code generated for this occurrence.
	Status int `json:"status"`

	// Title A short, human-readable summary of the problem.
	Title string `json:"title"`

	// Type A URI reference that identifies the problem type.
	Type string `json:"type"`
}

// RecordLoginJSONBody defines parameters for RecordLogin.
type RecordLoginJSONBody struct {
	LoginTime time.Time          `json:"login_time"`
	UserId    openapi_types.UUID `json:"user_id"`
}

// RecordLoginJSONRequestBody defines body for RecordLogin for application/json ContentType.
type RecordLoginJSONRequestBody RecordLoginJSONBody

// ServerInterface represents all server handlers.
type ServerInterface interface {

	// (GET /api/livez)
	Livez(w http.ResponseWriter, r *http.Request)

	// (GET /api/readyz)
	Readyz(w http.ResponseWriter, r *http.Request)

	// (GET /api/v1/analytics/users/daily/{day})
	GetDailyUniqueUsers(w http.ResponseWriter, r *http.Request, day openapi_types.Date)

	// (GET /api/v1/analytics/users/monthly/{month})
	GetMonthlyUniqueUsers(w http.ResponseWriter, r *http.Request, month string)

	// (POST /api/v1/logins)
	RecordLogin(w http.ResponseWriter, r *http.Request)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// (GET /api/livez)
func (_ Unimplemented) Livez(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/readyz)
func (_ Unimplemented) Readyz(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/v1/analytics/users/daily/{day})
func (_ Unimplemented) GetDailyUniqueUsers(w http.ResponseWriter, r *http.Request, day openapi_types.Date) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (GET /api/v1/analytics/users/monthly/{month})
func (_ Unimplemented) GetMonthlyUniqueUsers(w http.ResponseWriter, r *http.Request, month string) {
	w.WriteHeader(http.StatusNotImplemented)
}

// (POST /api/v1/logins)
func (_ Unimplemented) RecordLogin(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// Livez operation middleware
func (siw *ServerInterfaceWrapper) Livez(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Livez(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// Readyz operation middleware
func (siw *ServerInterfaceWrapper) Readyz(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.Readyz(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetDailyUniqueUsers operation middleware
func (siw *ServerInterfaceWrapper) GetDailyUniqueUsers(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "day" -------------
	var day openapi_types.Date

	err = runtime.BindStyledParameterWithOptions("simple", "day", chi.URLParam(r, "day"), &day, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "date"})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "day", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetDailyUniqueUsers(w, r, day)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetMonthlyUniqueUsers operation middleware
func (siw *ServerInterfaceWrapper) GetMonthlyUniqueUsers(w http.ResponseWriter, r *http.Request) {

	var err error
	_ = err

	// ------------- Path parameter "month" -------------
	var month string

	err = runtime.BindStyledParameterWithOptions("simple", "month", chi.URLParam(r, "month"), &month, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: ""})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "month", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetMonthlyUniqueUsers(w, r, month)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// RecordLogin operation middleware
func (siw *ServerInterfaceWrapper) RecordLogin(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.RecordLogin(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/livez", wrapper.Livez)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/readyz", wrapper.Readyz)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/analytics/users/daily/{day}", wrapper.GetDailyUniqueUsers)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/api/v1/analytics/users/monthly/{month}", wrapper.GetMonthlyUniqueUsers)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/api/v1/logins", wrapper.RecordLogin)
	})

	return r
}

type LivezRequestObject struct {
}

type LivezResponseObject interface {
	VisitLivezResponse(w http.ResponseWriter) error
}

type Livez200JSONResponse struct {
	Data *string `json:"data,omitempty"`
}

func (response Livez200JSONResponse) VisitLivezResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type ReadyzRequestObject struct {
}

type ReadyzResponseObject interface {
	VisitReadyzResponse(w http.ResponseWriter) error
}

type Readyz200JSONResponse struct {
	Data *string `json:"data,omitempty"`
}

func (response Readyz200JSONResponse) VisitReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type Readyz500JSONResponse ErrorResponse

func (response Readyz500JSONResponse) VisitReadyzResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type GetDailyUniqueUsersRequestObject struct {
	Day openapi_types.Date `json:"day"`
}

type GetDailyUniqueUsersResponseObject interface {
	VisitGetDailyUniqueUsersResponse(w http.ResponseWriter) error
}

type GetDailyUniqueUsers200JSONResponse struct {
	Data *int `json:"data,omitempty"`
}

func (response GetDailyUniqueUsers200JSONResponse) VisitGetDailyUniqueUsersResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetDailyUniqueUsers400JSONResponse ErrorResponse

func (response GetDailyUniqueUsers400JSONResponse) VisitGetDailyUniqueUsersResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type GetMonthlyUniqueUsersRequestObject struct {
	Month string `json:"month"`
}

type GetMonthlyUniqueUsersResponseObject interface {
	VisitGetMonthlyUniqueUsersResponse(w http.ResponseWriter) error
}

type GetMonthlyUniqueUsers200JSONResponse struct {
	Data *int `json:"data,omitempty"`
}

func (response GetMonthlyUniqueUsers200JSONResponse) VisitGetMonthlyUniqueUsersResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)
	_, err := buf.WriteTo(w)
	return err
}

type GetMonthlyUniqueUsers400JSONResponse ErrorResponse

func (response GetMonthlyUniqueUsers400JSONResponse) VisitGetMonthlyUniqueUsersResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type RecordLoginRequestObject struct {
	Body *RecordLoginJSONRequestBody
}

type RecordLoginResponseObject interface {
	VisitRecordLoginResponse(w http.ResponseWriter) error
}

type RecordLogin201Response struct {
}

func (response RecordLogin201Response) VisitRecordLoginResponse(w http.ResponseWriter) error {
	w.WriteHeader(201)
	return nil
}

type RecordLogin400JSONResponse ErrorResponse

func (response RecordLogin400JSONResponse) VisitRecordLoginResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {

	// (GET /api/livez)
	Livez(ctx context.Context, request LivezRequestObject) (LivezResponseObject, error)

	// (GET /api/readyz)
	Readyz(ctx context.Context, request ReadyzRequestObject) (ReadyzResponseObject, error)

	// (GET /api/v1/analytics/users/daily/{day})
	GetDailyUniqueUsers(ctx context.Context, request GetDailyUniqueUsersRequestObject) (GetDailyUniqueUsersResponseObject, error)

	// (GET /api/v1/analytics/users/monthly/{month})
	GetMonthlyUniqueUsers(ctx context.Context, request GetMonthlyUniqueUsersRequestObject) (GetMonthlyUniqueUsersResponseObject, error)

	// (POST /api/v1/logins)
	RecordLogin(ctx context.Context, request RecordLoginRequestObject) (RecordLoginResponseObject, error)
}

type StrictHandlerFunc func(ctx context.Context, w http.ResponseWriter, r *http.Request, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

type StrictHTTPServerOptions struct {
	RequestErrorHandlerFunc  func(w http.ResponseWriter, r *http.Request, err error)
	ResponseErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: StrictHTTPServerOptions{
		RequestErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		},
		ResponseErrorHandlerFunc: func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusInternalServerError)
		},
	}}
}

func NewStrictHandlerWithOptions(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc, options StrictHTTPServerOptions) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares, options: options}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
	options     StrictHTTPServerOptions
}

// Livez operation middleware
func (sh *strictHandler) Livez(w http.ResponseWriter, r *http.Request) {
	var request LivezRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.Livez(ctx, request.(LivezRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Livez")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(LivezResponseObject); ok {
		if err := validResponse.VisitLivezResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// Readyz operation middleware
func (sh *strictHandler) Readyz(w http.ResponseWriter, r *http.Request) {
	var request ReadyzRequestObject

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.Readyz(ctx, request.(ReadyzRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "Readyz")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(ReadyzResponseObject); ok {
		if err := validResponse.VisitReadyzResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetDailyUniqueUsers operation middleware
func (sh *strictHandler) GetDailyUniqueUsers(w http.ResponseWriter, r *http.Request, day openapi_types.Date) {
	var request GetDailyUniqueUsersRequestObject

	request.Day = day

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetDailyUniqueUsers(ctx, request.(GetDailyUniqueUsersRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetDailyUniqueUsers")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetDailyUniqueUsersResponseObject); ok {
		if err := validResponse.VisitGetDailyUniqueUsersResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// GetMonthlyUniqueUsers operation middleware
func (sh *strictHandler) GetMonthlyUniqueUsers(w http.ResponseWriter, r *http.Request, month string) {
	var request GetMonthlyUniqueUsersRequestObject

	request.Month = month

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.GetMonthlyUniqueUsers(ctx, request.(GetMonthlyUniqueUsersRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "GetMonthlyUniqueUsers")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(GetMonthlyUniqueUsersResponseObject); ok {
		if err := validResponse.VisitGetMonthlyUniqueUsersResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}

// RecordLogin operation middleware
func (sh *strictHandler) RecordLogin(w http.ResponseWriter, r *http.Request) {
	var request RecordLoginRequestObject

	var body RecordLoginJSONRequestBody
	if err := json.NewDecoder(r.Body).Decode(&body); err != nil {
		sh.options.RequestErrorHandlerFunc(w, r, fmt.Errorf("can't decode JSON body: %w", err))
		return
	}
	request.Body = &body

	handler := func(ctx context.Context, w http.ResponseWriter, r *http.Request, request interface{}) (interface{}, error) {
		return sh.ssi.RecordLogin(ctx, request.(RecordLoginRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RecordLogin")
	}

	response, err := handler(r.Context(), w, r, request)

	if err != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, err)
	} else if validResponse, ok := response.(RecordLoginResponseObject); ok {
		if err := validResponse.VisitRecordLoginResponse(w); err != nil {
			sh.options.ResponseErrorHandlerFunc(w, r, err)
		}
	} else if response != nil {
		sh.options.ResponseErrorHandlerFunc(w, r, fmt.Errorf("unexpected response type: %T", response))
	}
}